Add checkout logic

+++ mydeal/src/main/java/com/mydeal/domain/entities/Customer.java
@@ -37,7 +37,7 @@ public class Customer {
     private String password;
 
     @Column(name = "credit_limit", precision = 10, scale = 2)
-    private BigDecimal creditLimit;
+    private Double creditLimit;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "address_id")
