+++ mydeal/src/main/java/com/mydeal/presentation/utils/Utils.java
@@ -14,7 +14,7 @@ public class Utils {
         customer.setEmail(request.getParameter(RequestKey.RQ_CustomerEmail));
         customer.setPhoneNumber(request.getParameter(RequestKey.RQ_CustomerPhoneNumber));
         customer.setDateOfBirth(LocalDate.parse(request.getParameter(RequestKey.RQ_CustomerDOB)));
-        customer.setCreditLimit(BigDecimal.valueOf(Double.parseDouble(request.getParameter(RequestKey.RQ_CustomerCreditLimit))));
+        customer.setCreditLimit(Double.parseDouble(request.getParameter(RequestKey.RQ_CustomerCreditLimit)));
         Address address = (customer.getAddress() == null ? new Address() : customer.getAddress());
         address.setCity(request.getParameter(RequestKey.RQ_CustomerCity));
         address.setStreet(request.getParameter(RequestKey.RQ_CustomerStreet));
